requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Werkzeug
python-dotenv
email_validator
